fix(store): xmax 显式 cast bigint 修复 pgx xid 扫描错误
pgx 在 binary 协议下无法将 xid (OID 28) 扫描到 *int64，
CI 的 Postgres 18 触发此问题。RETURNING xmax::bigint 显式转型。

+++ store/sources.go
@@ -166,7 +166,7 @@ func (s *Store) UpsertSource(ctx context.Context, src *types.Source) (id int64, 
 		     title      = COALESCE(NULLIF(EXCLUDED.title, ''), sources.title),
 		     config     = EXCLUDED.config,
 		     updated_at = now()
-		 RETURNING id, xmax`,
+		 RETURNING id, xmax::bigint`,
 		src.Platform, src.Capability, src.URL, src.Title, cfg).Scan(&id, &xmax)
 	if err != nil {
 		return 0, false, types.NewAppError(types.CodeDatabase,
